+++ views/layouts/main.handlebars
@@ -2,7 +2,7 @@
 <html>
 	<head>
 		<meta charset="utf-8">
-		<title>Preguntados ODS</title>
+		<title>Tetris Game</title>
 		<link rel="stylesheet" href="/css/styles.css">
 		<script src="/js/dom.js"></script>
 		<meta name="viewport" content="width=device-width, initial-scale=1">
