+++ views/tetris.handlebars
@@ -1,15 +1,5 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Tetris Game</title>
-    <link rel="stylesheet" href="style.css">
-</head>
-<body>
-    <h2>Puntaje: 0</h2>
-    <canvas id="tetris" height="600" width="300"></canvas>
-</body>
-<script src="dom.js"></script>
-</html>
+
+<h2>Puntaje: 0</h2>
+<canvas id="tetris" height="600" width="300"></canvas>
+
+<script src="/js/tetris.js"></script>
